requirements.txt
flask==3.0.0
flask-cors==6.0.0
gunicorn==21.2.0
joblib==1.5.3
numpy==2.2.6
scikit-learn==1.7.2
xgboost>=1.6.0
matplotlib==3.10.8
pandas==2.3.3
nltk==3.9.4
requests==2.33.0
beautifulsoup4>=4.12.0
datasets
seaborn
newspaper3k
# Uncomment when re-enabling SHAP:
# shap>=0.41.0
lxml==4.9.4
